pyproject.toml
[project]
name = "chatbot-me"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.29.1",
    "openai>=1.78.1",
    "openai-agents>=0.0.15",
    "pypdf>=5.5.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
